refactor(preview): shift login context provider to preview
