+++ .storybook/preview.js
@@ -1,7 +1,9 @@
-import { ThemeProvider } from "@opengovsg/design-system-react"
+import { ThemeProvider, useToast } from "@opengovsg/design-system-react"
 import { initialize, mswDecorator } from "msw-storybook-addon"
 import { QueryClient, QueryClientProvider } from "react-query"
+import { LoginContext } from "contexts/LoginContext"
 import theme from "theme"
+import { MOCK_USER } from "mocks/constants"
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
@@ -18,6 +20,30 @@ export const parameters = {
 
 // Initialize MSW
 initialize()
+
+const withLoginContext = (Story) => {
+  const toast = useToast()
+  return (
+    <LoginContext.Provider
+      value={{
+        logout: async () => {
+          toast({
+            title: "User is logged out",
+            status: "success",
+            duration: 9000,
+            isClosable: true,
+          })
+        },
+        ...MOCK_USER,
+        verifyLoginAndSetLocalStorage: async () => {
+          return undefined
+        },
+      }}
+    >
+      <Story />
+    </LoginContext.Provider>
+  )
+}
 
 const withThemeProvider = (Story) => (
   <ThemeProvider resetCSS theme={theme}>
@@ -41,4 +67,9 @@ const withReactQuery = (Story) => {
   )
 }
 
-export const decorators = [withThemeProvider, withReactQuery, mswDecorator]
+export const decorators = [
+  withLoginContext,
+  withThemeProvider,
+  withReactQuery,
+  mswDecorator,
+]
